+++ src/DotNetDelice.Licensing/LicenseCache.fs
@@ -234,26 +234,24 @@ let private convertIfGitHub (licenseUrl : string) =
 let findMatchingLicense licenseContents =
     descriptions |> Map.tryFindKey (fun _ licenseTemplate -> diceCoefficient licenseTemplate licenseContents > 0.98)
 
-let checkLicenseContents checkLicenseContent packageName licenseUrl =
-    if not checkLicenseContent then None
-    else
-        match dynamicLicenseCache.TryFind licenseUrl with
-        | Some lc -> Some lc
-        | None ->
-            if failedUrls |> List.contains licenseUrl then None
-            else
-                try
-                    let licenseContents = convertIfGitHub licenseUrl |> Http.RequestString
-                    match findMatchingLicense licenseContents with
-                    | Some key ->
-                        let lc =
-                            { Expression = key
-                              Packages = Map.ofList [ (packageName, []) ] }
-                        dynamicLicenseCache <- dynamicLicenseCache.Add(licenseUrl, lc)
-                        Some lc
-                    | None ->
-                        failedUrls <- licenseUrl :: failedUrls
-                        None
-                with _ ->
+let checkLicenseContents packageName licenseUrl =
+    match dynamicLicenseCache.TryFind licenseUrl with
+    | Some lc -> Some lc
+    | None ->
+        if failedUrls |> List.contains licenseUrl then None
+        else
+            try
+                let licenseContents = convertIfGitHub licenseUrl |> Http.RequestString
+                match findMatchingLicense licenseContents with
+                | Some key ->
+                    let lc =
+                        { Expression = key
+                          Packages = Map.ofList [ (packageName, []) ] }
+                    dynamicLicenseCache <- dynamicLicenseCache.Add(licenseUrl, lc)
+                    Some lc
+                | None ->
                     failedUrls <- licenseUrl :: failedUrls
                     None
+            with _ ->
+                failedUrls <- licenseUrl :: failedUrls
+                None